PBI-2: Refactor - Folder structure - Exception name
