+++ UsersService/src/main/java/com/lenhatthanh/usersservice/config/SpringSecurity.java
@@ -1,4 +1,4 @@
-package com.lenhatthanh.usersservice.security;
+package com.lenhatthanh.usersservice.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
