+++ UsersService/src/main/java/com/lenhatthanh/usersservice/exception/EmailAlreadyExistException.java
@@ -1,0 +1,7 @@
+package com.lenhatthanh.usersservice.exception;
+
+public class EmailAlreadyExistException extends ApplicationException {
+    public EmailAlreadyExistException(String message) {
+        super(message);
+    }
+}
