--- UsersService/src/main/java/com/lenhatthanh/usersservice/exception/UserAlreadyExistException.java
@@ -1,7 +1,0 @@
-package com.lenhatthanh.usersservice.exception;
-
-public class UserAlreadyExistException extends ApplicationException {
-    public UserAlreadyExistException(String message) {
-        super(message);
-    }
-}
